requirements.txt
apache-airflow
python-dotenv
confluent-kafka
cassandra-driver
pandas